+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Fixed
+- Updated cached OpenAI model metadata.
+
+
 ## [v12.5.3] - 2024-04-05
 ### Fixed
 - The speed benchmark for OpenAI models was extremely slow, due to an issue with the